Add config option to set QOS of published messages
This commit adds a new config option to lpmqtt to set the QOS level
for messages published to MQTT. The QOS used in MQTT is the min value
between what the client and publisher use. By default lpmqtt will use
a QOS level of 0, but if operating in an environment where more
guarantees are needed on delivery you can set this to be higher.

Change-Id: If2ce821565551fdf686e253d06356080702c05a1