Add config option to set the imap idle timeout
This commit adds a new config option to the imap section, idle-timeout,
which is used to set the timeout duration for the idle call. By default
imaplib2 sets this to 29mins. This if far too long especially on an imap
server that doesn't have imap idle configured properly. What ends up
happening is that all the lp events get bunched together and pushed to
mqtt at once every 29mins. This isn't really useful as an event stream.
So this makes it configurable and decreases the default setting to 1 min
which seems like a much more useful default value, albeit at the cost of
more imap traffic to fetch messages.

Change-Id: I98fc9778f0adc548d28e7bbd8600f05ff4946ba2